Move IMDS checks into SSRF detector
